factor AddCasesModel to separate component
