+++ src/design-system/Modal.tsx
@@ -1,39 +1,33 @@
-import React, { useState } from "react";
+import React from "react";
 import styled from "styled-components";
 
-import Colors from "./Colors";
 import ModalDialog from "./ModalDialog";
 import { TitleProps } from "./ModalTitle";
 
 const ModalContainer = styled.div``;
 
-const ModalTrigger = styled.button`
-  color: ${Colors.green};
-  cursor: pointer;
-  font-family: "Libre Baskerville", serif;
-  font-size: 32px;
-  line-height: 32px;
-  letter-spacing: -0.03em;
-`;
-
-interface Props {
+export interface Props {
   modalTitle?: TitleProps["title"];
+  onClose?: () => void;
+  open: boolean;
+  setOpen: (open: boolean) => void;
   trigger?: string | React.ReactElement<any>;
-  children?: any;
 }
 
 const Modal: React.FC<Props> = (props) => {
-  const [open, setOpen] = useState(false);
-  const { trigger, modalTitle, children } = props;
+  const { trigger, modalTitle, children, open, setOpen } = props;
+
+  const closeModal = () => {
+    if (props.onClose) {
+      props.onClose();
+    }
+    setOpen(false);
+  };
 
   return (
     <ModalContainer>
-      <ModalTrigger onClick={() => setOpen(true)}>{trigger}</ModalTrigger>
-      <ModalDialog
-        title={modalTitle}
-        open={open}
-        closeModal={() => setOpen(false)}
-      >
+      <div onClick={() => setOpen(true)}> {trigger}</div>
+      <ModalDialog title={modalTitle} open={open} closeModal={closeModal}>
         {children}
       </ModalDialog>
     </ModalContainer>
