+++ src/page-multi-facility/FacilityRow.tsx
@@ -1,26 +1,17 @@
-import { startOfDay, startOfToday } from "date-fns";
 import { navigate } from "gatsby";
-import { pick } from "lodash";
 import React, { useContext, useState } from "react";
 import styled from "styled-components";
 
 import { saveFacility } from "../database/index";
 import Colors, { MarkColors as markColors } from "../design-system/Colors";
 import { DateMMMMdyyyy } from "../design-system/DateFormats";
-import InputButton from "../design-system/InputButton";
-import InputDate from "../design-system/InputDate";
 import InputDescription from "../design-system/InputDescription";
-import ModalDialog from "../design-system/ModalDialog";
 import { Spacer } from "../design-system/Spacer";
 import { useFlag } from "../feature-flags";
 import CurveChartContainer from "../impact-dashboard/CurveChartContainer";
-import {
-  EpidemicModelUpdate,
-  persistedKeys as facilityModelKeys,
-  totalConfirmedCases,
-} from "../impact-dashboard/EpidemicModelContext";
-import { AgeGroupGrid } from "../impact-dashboard/FacilityInformation";
+import { totalConfirmedCases } from "../impact-dashboard/EpidemicModelContext";
 import useModel from "../impact-dashboard/useModel";
+import AddCasesModal from "./AddCasesModal";
 import { FacilityContext } from "./FacilityContext";
 import {
   useChartDataFromProjectionData,
@@ -52,22 +43,6 @@ const CaseText = styled.div`
   color: ${Colors.darkRed};
 `;
 
-const ModalContents = styled.div`
-  align-items: flex-start;
-  display: flex;
-  flex-direction: column;
-  font-weight: normal;
-  justify-content: flex-start;
-  margin-top: 30px;
-`;
-
-const HorizRule = styled.div`
-  border-bottom: 0.5px solid ${Colors.darkGray};
-  padding-bottom: 20px;
-  margin-bottom: 20px;
-  width: 100%;
-`;
-
 // TODO: validate the arguments?
 const handleSubClick = (fn?: Function, ...args: any[]) => {
   return (event: React.MouseEvent<Element>) => {
@@ -84,29 +59,11 @@ interface Props {
   scenarioId: string;
 }
 
-// Create a diff of the model to store changes in the update cases modal.
-// This is necessary so that we don't update the current modal if the modal is thrown away w/o saving or
-// if the date added in the modal is prior to the current date (backfill)
-const useModelDiff = (): [
-  EpidemicModelUpdate,
-  (update: EpidemicModelUpdate) => void,
-  () => void,
-] => {
-  const [diff, setDiff] = useState({});
-  const mergeDiff = (update: EpidemicModelUpdate) => {
-    setDiff({ ...diff, ...update });
-  };
-  const resetDiff = () => {
-    setDiff({});
-  };
-  return [diff, mergeDiff, resetDiff];
-};
-
 const FacilityRow: React.FC<Props> = ({
   facility: initialFacility,
   scenarioId: scenarioId,
 }) => {
-  const [model, updateModel] = useModel();
+  const [model] = useModel();
 
   const { rtData, setFacility } = useContext(FacilityContext);
   const [facility, updateFacility] = useState(initialFacility);
@@ -127,56 +84,22 @@ const FacilityRow: React.FC<Props> = ({
     navigate("/facility");
   };
 
-  let [modelDiff, fakeUpdateModel, resetModelDiff] = useModelDiff();
-  const newModel = { ...model, ...modelDiff };
-
-  // Open/close update case counts modal. Saves data on clicking the save button, discards otherwise
-  const [caseCountsModal, updateCaseCountsModal] = useState(false);
-  const openCaseCountsModal = () => {
-    updateCaseCountsModal(true);
-  };
-  const closeCaseCountsModal = () => {
-    resetModelDiff();
-    updateCaseCountsModal(false);
-  };
-
-  const save = () => {
-    // Ensure that we don't insert keys (like `localeDataSource`) that is in model but not in the facility modelInputs
-    const modelInputs = {
-      ...facility.modelInputs,
-      ...pick(newModel, facilityModelKeys),
-    };
-    // Update the local state iff
-    // The observedAt date in the modal is more recent than the observedAt date in the current modelInputs.
-    // This needs to happen so that facility data will show the most updated data w/o requiring a hard reload.
-    if (
-      newModel.observedAt &&
-      model.observedAt &&
-      startOfDay(newModel.observedAt) >= startOfDay(model.observedAt)
-    ) {
-      updateFacility({ ...facility, modelInputs });
-      updateModel(modelDiff);
-    }
-    // Save to DB with model changes
-    saveFacility(scenarioId, {
-      id,
-      modelInputs,
-    });
-    closeCaseCountsModal();
-  };
-
   return (
     <>
       <div onClick={openFacilityPage} className="cursor-pointer">
         <DataContainer className="flex flex-row mb-8 border-b">
           <div className="w-2/5 flex flex-col justify-between">
             <div className="flex flex-row h-full">
-              <CaseText
+              <div
                 className="w-1/4 font-bold"
-                onClick={handleSubClick(openCaseCountsModal)}
+                onClick={(e) => e.stopPropagation()}
               >
-                {confirmedCases}
-              </CaseText>
+                <AddCasesModal
+                  facility={facility}
+                  trigger={<CaseText>{confirmedCases}</CaseText>}
+                  updateFacility={updateFacility}
+                />
+              </div>
               <FacilityNameLabel onClick={handleSubClick()}>
                 <InputDescription
                   description={name}
@@ -220,31 +143,6 @@ const FacilityRow: React.FC<Props> = ({
           </div>
         </DataContainer>
       </div>
-      <ModalDialog
-        closeModal={closeCaseCountsModal}
-        open={caseCountsModal}
-        title="Add Cases"
-      >
-        <ModalContents>
-          <InputDate
-            labelAbove={"Date observed"}
-            onValueChange={(date) => {
-              if (date) {
-                fakeUpdateModel({ observedAt: date });
-              }
-            }}
-            valueEntered={newModel.observedAt || startOfToday()}
-          />
-          <HorizRule />
-          <AgeGroupGrid
-            model={newModel}
-            updateModel={fakeUpdateModel}
-            collapsible={true}
-          />
-          <HorizRule />
-          <InputButton label="Save" onClick={save} />
-        </ModalContents>
-      </ModalDialog>
     </>
   );
 };
